requirements.txt
yt-dlp
google-genai
python-dotenv
spacy
keybert
networkx
matplotlib
pyvis
numpy
flask